refactor: refactored the search page
