+++ src/app/core/services/search/search.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
 import { BehaviorSubject } from 'rxjs';
 import { SearchState } from '../../../shared/models/search.model';
+import { ApiService } from '../../services/api/api.service';
 
 @Injectable({
   providedIn: 'root',
@@ -14,8 +15,14 @@ export class SearchService {
   });
 
   private payload: any = null;
+  private collectionsListSource = new BehaviorSubject<any[]>([]);
+  collectionsList$ = this.collectionsListSource.asObservable();
+
+
+  constructor(private apiService: ApiService) {}
 
   searchResult$ = this.searchStateSource.asObservable();
+
 
   updateSearchResult(data: any) {
     this.searchStateSource.next({
@@ -28,15 +35,13 @@ export class SearchService {
   }
 
   updateSearchQuery(payload: any) {
-    // Update search state with query, textType, and findable
     this.searchStateSource.next({
       ...this.searchStateSource.value,
       query: payload.search.query,
       textType: payload.search.textTypes.textType,
       findable: payload.search.findable,
     });
-  
-    // Update payload
+
     this.payload = payload;
   }
 
@@ -49,12 +54,40 @@ export class SearchService {
     });
   }
 
-  // Simplified payload methods
   getPayload(): any {
     return this.payload;
   }
 
   setPayload(newPayload: any): void {
     this.payload = newPayload;
+  }
+
+  fetchCollectionsList(): void {
+    this.apiService.getCollectionsList().subscribe({
+      next: (response) => {
+        if (response.ok) {
+          const parsedBody = JSON.parse(response.body);
+          if (parsedBody?.search?.valuefacets?.facet) {
+            const collections = parsedBody.search.valuefacets.facet.map((facet: any) => {
+              const items = Array.isArray(facet.item) ? facet.item : [facet.item].filter(Boolean);
+              return {
+                header: facet?.label,
+                displayType: facet?.displayType,
+                collectionTypes: items.map((item: any) => ({
+                  label: item?.label,
+                  selected: item?.selected,
+                  value: item?.value,
+                })),
+              };
+            });
+
+            this.collectionsListSource.next(collections);
+          }
+        }
+      },
+      error: (err) => {
+        console.error('Error fetching collections:', err);
+      },
+    });
   }
 }
