+++ src/app/features/search-find-content/search-find-content.component.html
@@ -95,9 +95,9 @@
           <div class="d-flex justify-content-between align-items-center">
             <!-- Pagination Info -->
             <ng-template #paginationInfo>
-              <p class="mb-0">
-                Showing {{ ((currentPage - 1) * resultsPerPage + 1) }}-{{ Math.min(currentPage * resultsPerPage, totalResults) }} of {{ totalResults }}
-              </p>
+                <p class="mb-0" [ngStyle]="{ visibility: (!loading && totalResults > 0) ? 'visible' : 'hidden' }">
+                  Showing {{ ((currentPage - 1) * resultsPerPage + 1) }}-{{ Math.min(currentPage * resultsPerPage, totalResults) }} of {{ totalResults }}
+                </p>
             </ng-template>
 
             <!-- Render the pagination info template -->
@@ -108,7 +108,7 @@
         
           <hec-search-results></hec-search-results>
         
-          @if (!loading) {
+          @if (!loading && totalResults > 0) {
             <div class="d-lg-flex justify-content-between align-items-center mt-2">
               <!-- Pagination Info (for larger screens) -->
               <ng-container *ngTemplateOutlet="paginationInfo"></ng-container>
