+++ src/app/shared/components/search-results/search-results.component.html
@@ -1,37 +1,42 @@
 <div class="loadingcontainer position-relative">
-    <!-- Display loading message while data is being fetched -->
-    @if (loading) {
-      <div class="hec-loader-content">
-      </div>
-    }
-  
-    <!-- Display search results if available -->
-    @if (!loading) {
-      @for (result of searchResults; track result.title) {
-        <div class="card mb-3 shadow-sm">
-          <div class="card-body gap-3 d-flex">
-            <img src="{{result.imageUrl}}" alt="">
-            <div>
-                <div class="d-flex align-items-center">
-                    <p class="badge bg-primary me-2 mb-0">{{ result.type }}</p>
-                    <a [routerLink]="'/preview'" [queryParams]="{ guid: result.guid }" routerLinkActive="router-link-active" class="text-decoration-none">{{ result.title }}</a>
-                </div>
-        
-                  <p class="card-text">
-                    <strong>By:</strong> {{ result.authors }}<br />
-                    <strong>&copy;{{ result.year }}</strong><br />
-                    <strong>Available Formats:</strong>
-                  </p>
-        
-                  <!-- Change button text based on data availability -->
-                  <button class="btn btn-outline-primary btn-sm">
-                    {{ result.enableAddButton ? 'Add Entire Book +' : 'Remove Book -' }}
-                  </button>
-            </div>
-           
+  <!-- Display loading message while data is being fetched -->
+  @if (loading) {
+    <div class="hec-loader-content">
+    </div>
+  }
+
+  <!-- Display message if no results are found -->
+  @if (!loading && searchResults.length === 0) {
+    <div class="alert alert-warning text-center">
+      No results are found.
+    </div>
+  }
+
+  <!-- Display search results if available -->
+  @if (!loading && searchResults.length > 0) {
+    @for (result of searchResults; track result.title) {
+      <div class="card mb-3 shadow-sm">
+        <div class="card-body gap-3 d-flex">
+          <img src="{{result.imageUrl}}" alt="">
+          <div>
+              <div class="d-flex align-items-center">
+                  <p class="badge bg-primary me-2 mb-0">{{ result.type }}</p>
+                  <a [routerLink]="'/preview'" [queryParams]="{ guid: result.guid }" routerLinkActive="router-link-active" class="text-decoration-none">{{ result.title }}</a>
+              </div>
+      
+                <p class="card-text">
+                  <strong>By:</strong> {{ result.authors }}<br />
+                  <strong>&copy;{{ result.year }}</strong><br />
+                  <strong>Available Formats:</strong>
+                </p>
+      
+                <!-- Change button text based on data availability -->
+                <button class="btn btn-outline-primary btn-sm">
+                  {{ result.enableAddButton ? 'Add Entire Book +' : 'Remove Book -' }}
+                </button>
           </div>
         </div>
-      }
+      </div>
     }
-  </div>
-  
+  }
+</div>
